requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.31.0
msgraph-sdk
azure-identity
azure-data-tables>=12.0.0
aiohttp
azure-ai-projects